UnOfficial gpSP kai 3.3 test 2
git-svn-id: http://svn.sourceforge.jp/svnroot/gpsp-kai/trunk/gpsp-kai-test@398 e750ed6d-7236-0410-a570-cc313d6b6496

+++ src/fbm_print.c
@@ -324,7 +324,7 @@ void *fbm_malloc(size_t size)
 
   if (size == 0) return NULL;
 
-  h_block = malloc(size);
+  h_block = (int)malloc(size);
 
   if (h_block < 0) return NULL;
